Move the project source directory
Change src.dir to make it point to the source dir according to the Gradle src structure.
The idea is to have a Gradle src structure and a hybrid build configuration to let the unit test work with Gradle on the GitHub Linux machines and a Netbeans src configuration which points to the Gradle src only for the actual code, not for the test code.