Refractor timeouts and errors
Implemention of a proper exception handling between threads. Exceptions launched in main function code are stored and then showed used in an assertion message. Other exceptions are all handled.

External interruptions are handled in all thread.

Proper implementation of the timeout message and of error handling. This is how the checks work:
- the test script is execly aligned with the main function execution: test OK
- one of the main function outputs is not aligned with the test script: test FAILED
- the main function code throws an uncaught exception: test FAILED
- the main function does not complete the test execution: test FAILED
- the main function does not execute before a timeout is reached: test FAILED
- the main function completes the test execution but it continues its execution: test OK if only outputs are done; test FAILED if any input is done because in this case the test script wouldn't contain the input data, so the main function read call would block until the timeout is reached.